Fix lib 'demo' script
